Rename MetadataStores
