--- src/Sulu/Bundle/AdminBundle/Resources/js/stores/MetadataStore/index.js
@@ -1,4 +1,0 @@
-// @flow
-import metadataStore from './MetadataStore';
-
-export default metadataStore;
