+++ src/Sulu/Bundle/AdminBundle/Resources/js/stores/metadataStore/index.js
@@ -1,0 +1,4 @@
+// @flow
+import metadataStore from './metadataStore';
+
+export default metadataStore;
