+++ src/Sulu/Bundle/AdminBundle/Resources/js/stores/metadataStore/metadataStore.js
